bravetogether_backend/requirements.txt
ansicon==1.89.0 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
attrs==21.4.0 ; python_version >= "3.9" and python_version < "4.0"
blessed==1.19.1 ; python_version >= "3.9" and python_version < "4.0"
botocore==1.29.54 ; python_version >= "3.9" and python_version < "4.0"
chalice==1.27.3 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
dnspython==2.3.0 ; python_version >= "3.9" and python_version < "4.0"
inquirer==2.10.1 ; python_version >= "3.9" and python_version < "4.0"
jinxed==1.2.0 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
jmespath==1.0.1 ; python_version >= "3.9" and python_version < "4.0"
pip==22.2.2 ; python_version >= "3.9" and python_version < "4.0"
pymongo==4.3.3 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
python-dotenv==0.21.1 ; python_version >= "3.9" and python_version < "4.0"
python-editor==1.0.4 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.9" and python_version < "4.0"
readchar==4.0.3 ; python_version >= "3.9" and python_version < "4.0"
setuptools==66.1.1 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.4.0 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.14 ; python_version >= "3.9" and python_version < "4.0"
wcwidth==0.2.6 ; python_version >= "3.9" and python_version < "4.0"
wheel==0.38.4 ; python_version >= "3.9" and python_version < "4.0"
